backend/requirements.txt
langchain
langchain-community
langchain-openai
langchain-core
langchain-text-splitters
langchain-community
langchain-openai
langchain-core
langchain-text-splitters
langchain-groq
langgraph
django
django-cors-headers
django-rest-framework

faiss-cpu
pypdf

jwt
django-allauth
dj-rest-auth